requirements.txt
fastapi==0.86.0
uvicorn[standard]==0.19.0
python-dotenv==0.21.0
sqlmodel==0.0.8